Split out config test
