+++ exporters/otlp/metrics/src/test/java/io/opentelemetry/exporter/otlp/metrics/OtlpGrpcMetricExporterTest.java
@@ -6,13 +6,9 @@
 package io.opentelemetry.exporter.otlp.metrics;
 
 import static io.opentelemetry.api.common.AttributeKey.stringKey;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatCode;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.exporter.otlp.internal.Marshaler;
-import io.opentelemetry.exporter.otlp.internal.grpc.OkHttpGrpcExporterBuilder;
 import io.opentelemetry.exporter.otlp.internal.metrics.ResourceMetricsMarshaler;
 import io.opentelemetry.exporter.otlp.testing.internal.AbstractGrpcTelemetryExporterTest;
 import io.opentelemetry.proto.metrics.v1.ResourceMetrics;
@@ -23,13 +19,10 @@ import io.opentelemetry.sdk.metrics.data.LongPointData;
 import io.opentelemetry.sdk.metrics.data.LongSumData;
 import io.opentelemetry.sdk.metrics.data.MetricData;
 import io.opentelemetry.sdk.resources.Resource;
-import java.net.URISyntaxException;
-import java.nio.charset.StandardCharsets;
 import java.time.Duration;
 import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
-import org.junit.jupiter.api.Test;
 
 class OtlpGrpcMetricExporterTest
     extends AbstractGrpcTelemetryExporterTest<MetricData, ResourceMetrics, OtlpGrpcMetricExporter> {
@@ -79,102 +72,5 @@ class OtlpGrpcMetricExporterTest
   @Override
   protected Marshaler[] toMarshalers(List<MetricData> telemetry) {
     return ResourceMetricsMarshaler.create(telemetry);
-  }
-
-  @Test
-  @SuppressWarnings("PreferJavaTimeOverload")
-  void validConfig() {
-    assertThatCode(() -> OtlpGrpcMetricExporter.builder().setTimeout(0, TimeUnit.MILLISECONDS))
-        .doesNotThrowAnyException();
-    assertThatCode(() -> OtlpGrpcMetricExporter.builder().setTimeout(Duration.ofMillis(0)))
-        .doesNotThrowAnyException();
-    assertThatCode(() -> OtlpGrpcMetricExporter.builder().setTimeout(10, TimeUnit.MILLISECONDS))
-        .doesNotThrowAnyException();
-    assertThatCode(() -> OtlpGrpcMetricExporter.builder().setTimeout(Duration.ofMillis(10)))
-        .doesNotThrowAnyException();
-
-    assertThatCode(() -> OtlpGrpcMetricExporter.builder().setEndpoint("http://localhost:4317"))
-        .doesNotThrowAnyException();
-    assertThatCode(() -> OtlpGrpcMetricExporter.builder().setEndpoint("http://localhost"))
-        .doesNotThrowAnyException();
-    assertThatCode(() -> OtlpGrpcMetricExporter.builder().setEndpoint("https://localhost"))
-        .doesNotThrowAnyException();
-    assertThatCode(() -> OtlpGrpcMetricExporter.builder().setEndpoint("http://foo:bar@localhost"))
-        .doesNotThrowAnyException();
-
-    assertThatCode(() -> OtlpGrpcMetricExporter.builder().setCompression("gzip"))
-        .doesNotThrowAnyException();
-    assertThatCode(() -> OtlpGrpcMetricExporter.builder().setCompression("none"))
-        .doesNotThrowAnyException();
-
-    assertThatCode(
-            () -> OtlpGrpcMetricExporter.builder().addHeader("foo", "bar").addHeader("baz", "qux"))
-        .doesNotThrowAnyException();
-
-    assertThatCode(
-            () ->
-                OtlpGrpcMetricExporter.builder()
-                    .setTrustedCertificates("foobar".getBytes(StandardCharsets.UTF_8)))
-        .doesNotThrowAnyException();
-
-    assertThatCode(
-            () ->
-                OtlpGrpcMetricExporter.builder()
-                    .setPreferredTemporality(AggregationTemporality.DELTA))
-        .doesNotThrowAnyException();
-    assertThat(
-            OtlpGrpcMetricExporter.builder()
-                .setPreferredTemporality(AggregationTemporality.DELTA)
-                .build()
-                .getPreferredTemporality())
-        .isEqualTo(AggregationTemporality.DELTA);
-    assertThat(OtlpGrpcMetricExporter.builder().build().getPreferredTemporality())
-        .isEqualTo(AggregationTemporality.CUMULATIVE);
-  }
-
-  @Test
-  @SuppressWarnings("PreferJavaTimeOverload")
-  void invalidConfig() {
-    assertThatThrownBy(() -> OtlpGrpcMetricExporter.builder().setTimeout(-1, TimeUnit.MILLISECONDS))
-        .isInstanceOf(IllegalArgumentException.class)
-        .hasMessage("timeout must be non-negative");
-    assertThatThrownBy(() -> OtlpGrpcMetricExporter.builder().setTimeout(1, null))
-        .isInstanceOf(NullPointerException.class)
-        .hasMessage("unit");
-    assertThatThrownBy(() -> OtlpGrpcMetricExporter.builder().setTimeout(null))
-        .isInstanceOf(NullPointerException.class)
-        .hasMessage("timeout");
-
-    assertThatThrownBy(() -> OtlpGrpcMetricExporter.builder().setEndpoint(null))
-        .isInstanceOf(NullPointerException.class)
-        .hasMessage("endpoint");
-    assertThatThrownBy(() -> OtlpGrpcMetricExporter.builder().setEndpoint("😺://localhost"))
-        .isInstanceOf(IllegalArgumentException.class)
-        .hasMessage("Invalid endpoint, must be a URL: 😺://localhost")
-        .hasCauseInstanceOf(URISyntaxException.class);
-    assertThatThrownBy(() -> OtlpGrpcMetricExporter.builder().setEndpoint("localhost"))
-        .isInstanceOf(IllegalArgumentException.class)
-        .hasMessage("Invalid endpoint, must start with http:// or https://: localhost");
-    assertThatThrownBy(() -> OtlpGrpcMetricExporter.builder().setEndpoint("gopher://localhost"))
-        .isInstanceOf(IllegalArgumentException.class)
-        .hasMessage("Invalid endpoint, must start with http:// or https://: gopher://localhost");
-
-    assertThatThrownBy(() -> OtlpGrpcMetricExporter.builder().setCompression(null))
-        .isInstanceOf(NullPointerException.class)
-        .hasMessage("compressionMethod");
-    assertThatThrownBy(() -> OtlpGrpcMetricExporter.builder().setCompression("foo"))
-        .isInstanceOf(IllegalArgumentException.class)
-        .hasMessage(
-            "Unsupported compression method. Supported compression methods include: gzip, none.");
-
-    assertThatThrownBy(() -> OtlpGrpcMetricExporter.builder().setPreferredTemporality(null))
-        .isInstanceOf(NullPointerException.class)
-        .hasMessage("preferredTemporality");
-  }
-
-  @Test
-  void usingOkHttp() {
-    assertThat(OtlpGrpcMetricExporter.builder().delegate)
-        .isInstanceOf(OkHttpGrpcExporterBuilder.class);
   }
 }
